Move group_by_dir to src/python/pants/util/dirutil.py
