+++ src/python/pants/backend/java/goals/tailor.py
@@ -18,7 +18,6 @@ from pants.core.goals.tailor import (
     PutativeTarget,
     PutativeTargets,
     PutativeTargetsRequest,
-    group_by_dir,
 )
 from pants.engine.fs import PathGlobs, Paths
 from pants.engine.internals.selectors import Get
@@ -26,6 +25,7 @@ from pants.engine.rules import collect_rules, rule
 from pants.engine.target import Target
 from pants.engine.unions import UnionRule
 from pants.source.filespec import FilespecMatcher
+from pants.util.dirutil import group_by_dir
 from pants.util.logging import LogLevel
