+++ src/python/pants/backend/scala/goals/tailor.py
@@ -20,7 +20,6 @@ from pants.core.goals.tailor import (
     PutativeTarget,
     PutativeTargets,
     PutativeTargetsRequest,
-    group_by_dir,
 )
 from pants.engine.fs import PathGlobs, Paths
 from pants.engine.internals.selectors import Get
@@ -28,6 +27,7 @@ from pants.engine.rules import collect_rules, rule
 from pants.engine.target import Target
 from pants.engine.unions import UnionRule
 from pants.source.filespec import FilespecMatcher
+from pants.util.dirutil import group_by_dir
 from pants.util.logging import LogLevel
